% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_empty_cols.R
\name{add_empty_cols}
\alias{add_empty_cols}
\title{Add empty columns to a data frame}
\usage{
add_empty_cols(.data, group = 1, n = 1)
}
\arguments{
\item{.data}{A data frame.}

\item{group}{The grouping after which blank columns will be added.  Entering a number will add blank columns after each numeric grouping of columns (i.e. enter 3 to insert blank columns after every third column).  Specify a quoted column name or vector of quoted column names to add blank columns after each identified column.  Defaults to 1.}

\item{n}{The number of empty rows to insert after each group.  Defaults to 1.}
}
\description{
Insert empty columns into a data frame by a specified group.
}
\examples{
add_empty_cols(mtcars[1:10, ], group = 3, n = 2)

add_empty_cols(mtcars[1:10,], group = c('am', 'gear'))
}
\keyword{blank}
\keyword{columns}
