% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_empty_rows.R
\name{add_empty_rows}
\alias{add_empty_rows}
\title{Add empty rows to a data frame}
\usage{
add_empty_rows(.data, group = 1, n = 1)
}
\arguments{
\item{.data}{A data frame.}

\item{group}{The grouping after which blank rows will be added.  Entering a number will add blank rows after each numeric grouping of rows (i.e. enter 3 to insert blank rows after every third row).  Specify a quoted column name or vector of quoted column names to group the data frame and add blank rows after each grouping.  Defaults to 1.}

\item{n}{The number of empty rows to insert after each group.  Defaults to 1.}
}
\description{
Insert empty rows into a data frame by a specified group.
}
\examples{
add_empty_rows(mtcars[1:10,], 3, 2)

add_empty_rows(mtcars[1:10,], group = c('am', 'gear'))
}
\keyword{blank}
\keyword{rows}
