% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_stack.R
\name{df_stack}
\alias{df_stack}
\title{Stack data frames}
\usage{
df_stack(.data, n = 1)
}
\arguments{
\item{.data}{A collection of data frames wrapped in a list. Use list(df1, df2, etc.).}

\item{n}{The number of rows to put between stacked data frames.  Defaults to 1.}
}
\description{
Stack data frames on top of each other to form one tall data frame.
}
\examples{
df1 <- head(mtcars, 10)
df2 <- head(iris, 8)
df_stack(list(df1, df2), n = 2)
}
\keyword{stack}
