% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toddler_functions.R
\name{.split_camel_case}
\alias{.split_camel_case}
\title{Helper function to split camelCase into words (internal)}
\usage{
.split_camel_case(x)
}
\arguments{
\item{x}{A character string that might be camelCase}
}
\value{
A string with spaces between words
}
\description{
Helper function to split camelCase into words (internal)
}
\keyword{internal}
