% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toddler_functions.R
\name{.to_title_case}
\alias{.to_title_case}
\title{Helper function for title case (internal)}
\usage{
.to_title_case(x)
}
\arguments{
\item{x}{A character string}
}
\value{
A title case string
}
\description{
Helper function for title case (internal)
}
\keyword{internal}
